review(pr2): revue de code feature/notes-crud — corrections appliquées
Problèmes détectés et corrigés:
1. evaluations.routes.js: ajout .toInt() sur numero, coefficient, classe_id, matiere_id, trimestre
   - Forçage de la conversion entière pour éviter les valeurs string/float
2. absences.service.js: validation explicite du trimestre dans getAbsencesByEleve
   - Vérification que trimestre ∈ [1, 2, 3] avant accès à l'objet moisParTrimestre
   - Évite tout comportement inattendu avec des valeurs hors plage
3. utilisateurs.routes.js: ajout limite par défaut (100) sur GET /contacts
   - Évite le retour de tous les utilisateurs sans limite pour les grands établissements
   - Paramètre ?limit= configurable par le client
Points positifs: transactions ACID, UPSERT, vérification droits, pagination, flags absences
REVIEW_PR2.md: rapport complet de la revue

+++ backend/src/modules/notes/evaluations.routes.js
@@ -14,12 +14,12 @@ router.post(
   requireRole('enseignant'),
   [
     body('type').isIn(['DS','Interrogation','Examen']).withMessage('Type invalide'),
-    body('numero').isInt({ min: 1 }).withMessage('Numéro invalide'),
-    body('coefficient').isInt({ min: 1 }).withMessage('Coefficient invalide'),
+    body('numero').isInt({ min: 1 }).toInt().withMessage('Numéro invalide'),
+    body('coefficient').isInt({ min: 1 }).toInt().withMessage('Coefficient invalide'),
     body('date').isISO8601().withMessage('Date invalide'),
-    body('classe_id').isInt({ min: 1 }).withMessage('Classe requise'),
-    body('matiere_id').isInt({ min: 1 }).withMessage('Matière requise'),
-    body('trimestre').isInt({ min: 1, max: 3 }).withMessage('Trimestre invalide'),
+    body('classe_id').isInt({ min: 1 }).toInt().withMessage('Classe requise'),
+    body('matiere_id').isInt({ min: 1 }).toInt().withMessage('Matière requise'),
+    body('trimestre').isInt({ min: 1, max: 3 }).toInt().withMessage('Trimestre invalide'),
   ],
   validate,
   async (req, res, next) => {
